HTTP Requests: async task separated from activities, less redundant code; created application for global context
