+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,7 @@
     <uses-feature android:name="android.hardware.location.gps" />
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
