+++ app/src/main/java/com/example/android/letseat/App.java
@@ -1,0 +1,19 @@
+package com.example.android.letseat;
+
+import android.app.Application;
+import android.content.Context;
+
+public class App extends Application {
+
+    private static Context mContext;
+
+    @Override
+    public void onCreate() {
+        super.onCreate();
+        mContext = this;
+    }
+
+    public static Context getContext() {
+        return mContext;
+    }
+}
